+++ src/models/payload_trait.rs
@@ -2,32 +2,32 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Temperature {
-    pub value: f32,
+    pub value: f64,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Humidity {
-    pub value: f32,
+    pub value: f64,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Light {
-    pub value: f32,
+    pub value: f64,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct AirPressure {
-    pub value: f32,
+    pub value: f64,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Motion {
-    pub value: i32,
+    pub value: i64,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct AirQuality {
-    pub value: i32,
+    pub value: i64,
 }
 
 pub trait PayloadTrait {}
